build&install commands for all runtimes
