+++ runtimes/cpp-17/CMakeLists.txt.fallback
@@ -1,0 +1,1 @@
+target_link_libraries(${PROJECT_NAME} PUBLIC curl)
