--- runtimes/dotnet-6.0/Example/Function.csproj
@@ -1,12 +1,0 @@
-<!-- All blocks are optional, only PackageReferences will be read. Additional properties only provide IDE support. -->
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <OutputType>Exe</OutputType>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
-  </ItemGroup>
-</Project>
