+++ runtimes/dotnet-6.0/Function.csproj.callback
@@ -1,0 +1,11 @@
+<!-- All blocks are optional, only PackageReferences will be read. Additional properties only provide IDE support. -->
+<Project Sdk="Microsoft.NET.Sdk.Web">
+
+  <PropertyGroup>
+    <TargetFramework>net6.0</TargetFramework>
+    <OutputType>Exe</OutputType>
+  </PropertyGroup>
+
+  <ItemGroup>
+  </ItemGroup>
+</Project>
