+++ runtimes/java-17.0/deps.gradle.fallback
@@ -1,0 +1,2 @@
+dependencies {
+}
