+++ runtimes/kotlin-1.6/deps.gradle.fallback
@@ -1,0 +1,2 @@
+dependencies {
+}
