+++ runtimes/node-14.5/build.sh
@@ -8,13 +8,21 @@ cp -R /usr/code/* /usr/builds
 
 # Install User Function Dependencies if package.json exists
 cd /usr/builds
-if [ -f package.json ]; then
-  npm install
+
+# Ensure package.json exists
+if [[ ! -f "package.json" ]]; then
+    mv /usr/local/src/package.json.fallback /usr/builds/package.json
 fi
 
-mkdir -p node_modules
+INSTALL_COMMAND=${1:-'npm install'}
+BUILD_COMMAND=${2:-''}    
+
+eval "$INSTALL_COMMAND"
+eval "$BUILD_COMMAND"
 
 # Merge the node_modules from the server into the user's node_modules to be restored later.
+cd /usr/builds
+mkdir -p node_modules
 cp -R /usr/local/src/node_modules/* /usr/builds/node_modules
 
 # Finish build by preparing tar to use for starting the runtime
