+++ runtimes/ruby-3.1/Gemfile.fallback
@@ -1,0 +1,1 @@
+source "https://rubygems.org"
