+++ runtimes/swift-5.5/Package.swift.fallback
@@ -1,0 +1,13 @@
+// swift-tools-version:5.5
+import PackageDescription
+
+let package = Package(
+    name: "swift-runtime",
+    dependencies: [],
+    targets: [
+        .executableTarget(
+            name: "swift-function",
+            dependencies: []
+        )
+    ]
+)
